Added calls to a logger when you are recording.
The triangle and square key on the controller are assigned to the starting and stopping of recording. When we are recording we pass the joystick inputs and the wheel speeds to the logger, the logger logs these as tabular data in a csv file. We also pass the current view of the robot, this gets saved as seperate jpg images.